PhysicsTransform追加＆微修正
PhysicsTransformStorageを追加した。(APIに生成も追加)
ID系の変数の型を変更した
APIからWorldStorageを消した。

+++ SoftBodyProject/Source/Physics/Storage/Body/RigidBodyStorage.cpp
@@ -158,6 +158,6 @@ BodyID RigidBodyStorage::GenerateBodyID(size_t _denseIndex, EntityID _ownerEntit
 		slots[index].ownerEntity = _ownerEntity;
 
 		// IDを作成(初代判定で1)
-		return BodyID{ index,slots[index].generation };
+		return BodyID{ (uint32_t)(index),slots[index].generation };
 	}
 }
